feat(lotto): 당첨 통계 및 수익률 계산 기능 구현
- 당첨 통계 출력 시작 안내 메시지 출력 기능 구현
- 등수별 당첨 결과 출력 기능 구현
- 총 수익 계산 기능 구현
- 수익률 계산 및 출력 기능 구현
- 해당하는 단위 테스트 구현 후 테스트 완료

+++ src/main/java/lotto/domain/Result.java
@@ -24,10 +24,10 @@ public class Result {
     return rankCounts;
   }
 
-  public int calculateTotalPrize() {
-    int total = 0;
+  public long calculateTotalPrize() {
+    long total = 0;
     for (Rank rank : Rank.values()) {
-      total += rank.getPrize() * rankCounts.get(rank);
+      total += (long) rank.getPrize() * rankCounts.get(rank);
     }
     return total;
   }
